del analogkeypad and renewed the dep

+++ dep/oem/brogue-rpg/backend/models/buff.py
@@ -1,6 +1,6 @@
 from .affix import AffixGroup
 
 class Buff:
-    def __init__(self, duration: int):
+    def __init__(self, duration: float):
         self.affixes = AffixGroup()
         self.duration = duration
